Fix overly-strong ASSERT in WeightTable::SetStartEnd()
A zero-length range at zero represented by [0, -1]  will be fine so long
as we don't actually try to reference the range. Stop short of switching
over to the simpler [0, 0) representation. Note that `weight_count` is
bounded by earlier size checks and fits in an int.

-- Add another ASSERT to earlier tests.

Bug: chromium:1219974
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/82155
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fxge/dib/cstretchengine.h
@@ -38,7 +38,7 @@ class CStretchEngine {
     static size_t TotalBytesForWeightCount(size_t weight_count);
 
     void SetStartEnd(int src_start, int src_end, size_t weight_count) {
-      CHECK_LT(static_cast<size_t>(src_end - src_start), weight_count);
+      CHECK_LT(src_end - src_start, static_cast<int>(weight_count));
       m_SrcStart = src_start;
       m_SrcEnd = src_end;
     }
@@ -61,7 +61,7 @@ class CStretchEngine {
     }
 
     int m_SrcStart;
-    int m_SrcEnd;           // Note: inclusive.
+    int m_SrcEnd;           // Note: inclusive, [0, -1] for empty range at 0.
     uint32_t m_Weights[1];  // Not really 1, variable size.
   };
